requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.4
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
